[PS][Hotfix]
arreglo de las migraciones

+++ database/migrations/2020_02_20_150610_alter_plate_table.php
@@ -14,7 +14,7 @@ class AlterPlateTable extends Migration
     public function up()
     {
         Schema::table('plate', function (Blueprint $table) {
-            $table->unsignedInteger('menu_id');
+            $table->unsignedBigInteger('menu_id');
             $table->foreign('menu_id')->references('id')->on('menu');
         });
     }
